warning: LF will be replaced by CRLF in front/src/pages/Main/style.module.css.
