+++ front/src/components/Navbar/RightMenu/style.module.css
@@ -1,0 +1,78 @@
+/* 오른쪽 메뉴 */
+.right_menu{display: flex; align-items: center;}
+.menu_icon{margin-right: 10px;}
+.menu_icon ul{display: flex;}
+.menu_icon ul li{width: 40px;height: 40px; display: flex; justify-content: center; align-items: center; border-radius: 50%; cursor: pointer; position: relative;}
+.menu_icon ul li:hover{background: #eee;}
+.menu_icon ul li .icon{font-size: 20px; color: #666;}
+.menu_icon ul li em{width: 100%; background: #555; font-size: 12px; color: #fff; visibility: hidden; opacity: 0; transition: all 0.4s; position: absolute;top: 110%;left: 50%; transform: translateX(-50%); text-align: center; padding: 5px 0;}
+.menu_icon ul li:hover em{visibility: visible; opacity: 1;}
+.menu_icon ul li:nth-child(3) em{width: 150%;}
+
+.date_list{position: relative; }
+.date_btt{height: 35px; padding: 0 10px; display: flex; justify-content: center; align-items: center; border: 1px solid #ccc; box-sizing: border-box; border-radius: 5px; cursor: pointer;}
+.date_btt:hover{background: #eee;}
+.date_btt strong{font-weight: 400;}
+.date_btt .icon{margin-left: 10px; color: #555;}
+
+.date_list .option_list{visibility: hidden; opacity: 0; width: 0; height: 0px; position: absolute; top: 120%;left: 0; padding: 15px 0px; background: #fff; box-sizing: border-box; box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px; z-index: 5555;}
+.date_list.active .option_list{visibility: visible; opacity: 1;  width: 160px; height: initial; transition: opacity,width,height 0.5s,0.1s;}
+.date_list .option_list > ul{}
+.date_list .option_list > ul > li{height: 40px;display: flex; justify-content: space-between; align-items: center; padding: 0 15px; cursor: pointer; }
+.date_list .option_list > ul > li:hover{background: #eee;}
+.date_list .option_list > ul > li > strong{font-size: 14px;font-weight: 400;}
+.date_list .option_list > ul > li > em{font-size: 14px;font-weight: 400; color: #555;}
+
+.line{width: 100%; height: 1px; display: block; background: #ccc;}
+
+.option_check{}
+.option_check ul{}
+.option_check ul li{height: 40px; display: flex; align-items: center; padding: 0 15px; cursor: pointer; }
+.option_check ul li:hover{background: #eee;}
+.option_check ul li .icon{margin-right: 10px;}
+
+.more{width: 40px;height: 40px; margin-left: 30px; display: flex; justify-content: center; align-items: center; border-radius: 50%; cursor: pointer; position: relative;}
+.more:hover{background: #eee;}
+.more .icon{font-size: 20px;}
+.more em{width: 170%; background: #555; font-size: 12px; color: #fff; visibility: hidden; opacity: 0; transition: all 0.4s; position: absolute;top: 110%;left: 50%; transform: translateX(-50%); text-align: center; padding: 5px 0;}
+.more:hover em{visibility: visible; opacity: 1;}
+
+.user{width: 32px;height: 32px; margin-left: 10px; border-radius: 50%; background: #ccc;  position: relative;}
+.user img{width: 100%; height: 100%; object-fit: cover;  border-radius: 50%; cursor: pointer;}
+.user > em{width: 250%; background: #555; font-size: 12px; color: #fff; visibility: hidden; opacity: 0; transition: all 0.4s; position: absolute;top: 125%;right: 3px; text-align: center; padding: 5px 0;}
+.user:hover > em{visibility: visible; opacity: 1;}
+
+.user_inpo{width: 360px; visibility: hidden; opacity: 0; background: #fff; position: absolute; top: 120%;right: 0; border-radius: 10px; box-shadow: rgba(14, 30, 37, 0.12) 0px 2px 4px 0px, rgba(14, 30, 37, 0.32) 0px 2px 16px 0px;}
+.user.active .user_inpo{visibility: visible; opacity: 1;}
+.user_profile{height: 230px; background: #fff; display: flex; flex-direction: column; justify-content: center; align-items: center; border-radius: 10px 10px 0 0;}
+.user_profile .user_img{width: 100px;height: 100px;background: #555; border-radius: 50%;}
+.user_profile h2{margin-top: 10px;font-size: 16px; font-weight: 400;}
+.user_profile em{font-size: 14px; color: #555;}
+.user_profile button{padding: 5px 20px; border-radius: 30px; border: 1px solid #ccc; margin-top: 10px; background: #fff; cursor: pointer;}
+.user_profile button:hover{background: #eee;}
+
+.account{border-top: 1px solid #ccc; border-bottom: 1px solid #ccc;}
+.account ul{}
+.account ul li{width: 100%;height: 50px; display: flex; align-items: center; padding: 0 20px; box-sizing: border-box; cursor: pointer;}
+.account ul li:hover{background: #eee;}
+.account ul li .account_img{width: 30px;height: 30px; background: #ccc; border-radius: 50%;}
+.account ul li .account_text{margin-left: 10px;}
+.account ul li .account_text p{font-size: 14px; line-height: 14px;}
+.account ul li .account_text p:last-child{color: #555;}
+
+.user_add{width: 100%;height: 40px; display: flex; align-items: center; padding: 0 20px; box-sizing: border-box; cursor: pointer;}
+.user_add:hover{background: #eee;}
+.user_add .icon{font-size: 20px; margin-left: 5px;}
+.user_add p{ margin-left: 10px; font-size: 16px;}
+
+.all_logout{width: 100%;height: 70px;display: flex; justify-content: center; align-items: center; border-bottom: 1px solid #ccc;}
+.all_logout button{font-size: 14px; padding: 7px 20px; background: #fff; border: 1px solid #ccc; border-radius: 5px; box-sizing: border-box; letter-spacing: -0.1em; cursor: pointer;}
+.all_logout button:hover{background: #eee;}
+
+.service{width: 100%;height: 60px;display: flex; align-items: center;}
+.service ul{width: 100%; display: flex; justify-content: center;}
+.service ul li{ padding: 0 5px; cursor: pointer;}
+.service ul li:hover{background: #eee;}
+.service ul li:nth-child(2){cursor: default;}
+.service ul li:nth-child(2):hover{background: transparent;}
+.service ul li em{}
